#23 Add test for radex module
